Beta_realise: v0.0.1

+++ cw/src/Image/canvas.c
@@ -11,7 +11,7 @@
  * @brief Главная функция, которая оперирует вспомогательными функциями, и рисует на изображении.
  * 
  * Используется в функции main.
- * @param Структура фигуры.
+ * @param figure фигура.
  */
 int canvas_logic(object_t figure) {
     BitmapFileHeader bmfh;
@@ -26,9 +26,9 @@ int canvas_logic(object_t figure) {
     Rgb color_fill = (Rgb){figure.color_fill_b, figure.color_fill_g, figure.color_fill_r};
     Rgb color =  (Rgb){figure.color_b, figure.color_g, figure.color_r};
 
-    figure.y_center = bmih.height - figure.y_center;
-    figure.y_left_up = bmih.height - figure.y_left_up;
-    figure.y_right_down = bmih.height - figure.y_right_down;
+    figure.y_center = (int)bmih.height - figure.y_center;
+    figure.y_left_up = (int)bmih.height - figure.y_left_up;
+    figure.y_right_down = (int)bmih.height - figure.y_right_down;
 
     switch (figure.mod) {
         case rect:
